+++ src/compojure/api/meta.clj
@@ -2,6 +2,7 @@
   (:require [compojure.api.common :refer :all]
             [compojure.core :refer [routes]]
             [ring.util.response :as response]
+            [ring.util.http-response :as http-response]
             [plumbing.core :refer :all]
             [plumbing.fnk.impl :as fnk-impl]
             [ring.swagger.core :as swagger]
@@ -23,12 +24,15 @@
    (fnk-impl/letk-input-schema-and-body-form
      nil (with-meta bind {:schema s/Any}) [] nil)))
 
-(defn body-coercer-middleware [handler model]
+(defn body-coercer-middleware [handler responses]
   (fn [request]
     (if-let [response (handler request)]
-      (assoc response
-        ::serializable? true
-        :body (schema/coerce! model (:body response))))))
+      (let [status (:status response)]
+        (if-let [model (responses status)]
+          (assoc response
+            ::serializable? true
+            :body (schema/coerce! model (:body response)))
+          response)))))
 
 (defn src-coerce!
   "Return source code for coerce! for a schema with coercer type,
@@ -39,6 +43,17 @@
      (keywordize-keys
        (~key ~+compojure-api-request+))
      ~type))
+
+;;
+;; Response messages mangling
+;;
+
+(defn- responses->messages [responses]
+  (for [[code model] responses
+        :when (not= code 200)]
+    {:code code
+     :message (or (some-> model meta :message) "")
+     :responseModel (eval model)}))
 
 ;;
 ;; Extension point
@@ -65,12 +80,26 @@
 ;;
 ;; Smart restructurings
 ;;
-
 (defmethod restructure-param :return [k model acc]
-  "Defines a return type and coerced the return value of a body against it."
+  "Defines a return type and coerced the return value of a body against it.
+   Examples:
+   :return MyModel
+   :return {:value String}
+   :return #{{:key (s/maybe Long)}}"
   (-> acc
       (update-in [:parameters] assoc k model)
-      (update-in [:middlewares] conj `(body-coercer-middleware ~model))))
+      (update-in [:responses] assoc 200 model)))
+
+(defmethod restructure-param :responses [_ responses acc]
+  "value is a map of http-response-code -> Model. Translates to both swagger
+   parameters and return model coercion. Models can be decorated with meta-data.
+   Examples:
+   :responses {403 ErrorEnvelope}
+   :responses {403 ^{:message \"Underflow\"} ErrorEnvelope}"
+  (let [messages (responses->messages responses)]
+    (-> acc
+        (update-in [:parameters :responseMessages] (comp distinct concat) messages)
+        (update-in [:responses] merge responses))))
 
 (defmethod restructure-param :body [_ [value model model-meta] acc]
   "reads body-params into a enchanced let. First parameter is the let symbol,
@@ -158,22 +187,25 @@
 (defn restructure [method [path arg & args]]
   (let [method-symbol (symbol (str (-> method meta :ns) "/" (-> method meta :name)))
         [parameters body] (extract-parameters args)
-        [lets letks middlewares] [[] [] nil]
+        [lets letks responses middlewares] [[] [] nil nil]
         [lets arg-with-request] (destructure-compojure-api-request lets arg)
         {:keys [lets
                 letks
+                responses
                 middlewares
                 parameters
                 body]} (reduce
-                         (fn [{:keys [lets letks middlewares parameters body]} [k v]]
+                         (fn [{:keys [lets letks responses middlewares parameters body]} [k v]]
                            (let [parameters (dissoc parameters k)
-                                 acc (map-of lets letks middlewares parameters body)]
+                                 acc (map-of lets letks responses middlewares parameters body)]
                              (restructure-param k v acc)))
-                         (map-of lets letks middlewares parameters body)
+                         (map-of lets letks responses middlewares parameters body)
                          parameters)
         body `(do ~@body)
         body (if (seq letks) `(letk ~letks ~body) body)
         body (if (seq lets) `(let ~lets ~body) body)
         body (if (seq parameters) `(meta-container ~parameters ~body) body)
-        body `(~method-symbol ~path ~arg-with-request ~body)]
-    (if (seq middlewares) `(middlewares [~@middlewares] ~body) body)))
+        body `(~method-symbol ~path ~arg-with-request ~body)
+        body (if (seq middlewares) `(middlewares [~@middlewares] ~body) body)
+        body (if responses `(body-coercer-middleware ~body  ~responses) body)]
+    body))